[KICKS] Refactor BE code
